API Requests in High Frequency
This is a code where I have been able to automate an API request that gave me a series of Dictionaries with crypto currencies, each for one coin. I get the information on mining difficulty, prices, time to mine blocks, etc, for dozens of coins and organize it one single panda dataframe of one line for each request. Then I append this to my historical data spreadsheet.